requirements.txt
Django==1.8.6
Jinja2==2.8
MarkupSafe==0.23
Pygments==2.0.2
argparse==1.2.1
backports-abc==0.4
backports.ssl-match-hostname==3.4.0.2
certifi==2015.9.6.2
decorator==4.0.4
django-jet==0.1.3
feedparser==5.2.1
functools32==3.2.3-2
google-api-python-client==1.4.2
httplib2==0.9.2
ipykernel==4.1.1
ipython==4.0.0
ipython-genutils==0.1.0
jsonschema==2.5.1
jupyter-client==4.1.1
jupyter-core==4.0.6
mistune==0.7.1
nbconvert==4.1.0
nbformat==4.0.1
notebook==4.0.6
oauth2client==1.5.1
path.py==8.1.2
pexpect==4.0.1
pickleshare==0.5
ptyprocess==0.5
pyasn1==0.1.9
pyasn1-modules==0.0.8
pytz==2015.7
pyzmq==15.0.0
requests==2.8.1
rsa==3.2.3
simplegeneric==0.8.1
simplejson==3.8.1
singledispatch==3.4.0.3
six==1.10.0
terminado==0.5
tornado==4.3
traitlets==4.0.0
uritemplate==0.6
wsgiref==0.1.2
xmltodict==0.9.2
